Regular Expression patterns for C++ syntax
the file for the regular expression patterns and substitutions for C++ Syntax, which will be used in sed utility for replacement.

Each line would contain a pair of pattern and substitution. No spaces before and after each line, otherwise they will be treated as part of pattern

Pattern and substitution should be seperated by tabs and/or blank spaces.

Any lines begin with # will be ignored as comment
Any empty lines which only contains tabs and/or blank spaces will be ignored